implementing changes to security config.

+++ src/main/java/com/ecomerceApi/Priscila/model/Role.java
@@ -5,8 +5,9 @@ import jakarta.persistence.*;
 @Entity
 @Table(name = "roles")
 public class Role {
-    @Id@GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Integer id;
 
     @Enumerated(EnumType.STRING)
     @Column
@@ -20,11 +21,11 @@ public class Role {
         this.name = name;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
